add detail package and move
