+++ app/src/main/java/com/muharuto/introduceclubactivities/detail/EpoxyConfig.java
@@ -1,7 +1,8 @@
-package com.muharuto.introduceclubactivities;
+package com.muharuto.introduceclubactivities.detail;
 
 import com.airbnb.epoxy.EpoxyDataBindingLayouts;
 import com.airbnb.epoxy.EpoxyDataBindingPattern;
+import com.muharuto.introduceclubactivities.R;
 
 @EpoxyDataBindingPattern(rClass = R.class, layoutPrefix = "epoxy_layout")
 @EpoxyDataBindingLayouts({
