+++ app/src/main/java/com/muharuto/introduceclubactivities/home/HomeFragment.kt
@@ -6,7 +6,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.muharuto.introduceclubactivities.ClubViewModel
+import com.muharuto.introduceclubactivities.detail.ClubViewModel
 import com.muharuto.introduceclubactivities.R
 import com.muharuto.introduceclubactivities.databinding.FragmentHomeBinding
